refactor(model): mechanically split model.rs into bounded modules (wave13 step2)
