+++ crates/assay-core/src/model/mod.rs
@@ -1,0 +1,8 @@
+mod serde;
+mod types;
+mod validation;
+
+pub use types::*;
+
+#[cfg(test)]
+mod tests;
